+++ hrdc/src/pages/directory.js
@@ -1,6 +1,6 @@
 import Image from "next/image";
 import SearchResults from "./api/searchresults";
-import Navbar from "../components/ui/navbar";
+import Navbar from "../components/Navbar";
 
 export default function Directory() {
     const handleOpenFile = (fileUrl) => {
